requirements.txt
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Cython==0.29.20
fastapi==0.58.0
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
h11==0.9.0
httplib2==0.18.1
idna==2.9
isort==4.3.21
jws==0.1.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2client==4.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
pydantic==1.5.1
pylint==2.5.3
Pyrebase4==4.3.0
python-jwt==2.0.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
starlette==0.13.4
toml==0.10.1
urllib3==1.25.9
uvicorn==0.11.5
websockets==8.1
wrapt==1.12.1
